se cargaron imagenes a productos, se agregaron controles en brandcontroller, se crearon comentarios y se corrigieron detalles de los mismos

+++ controller/BrandController.php
@@ -18,9 +18,14 @@ class BrandController
     public function delete($id)
     {
         try {
-            if ($id)
-            $this->brandModel->deleteBrand($id);
-            header("Location: ".BASE_URL  . "admins");
+            session_start();     
+            if (isset($_SESSION['email']) && $_SESSION['role'] == UserRole::ADMIN)  {
+                if ($id)
+                    $this->brandModel->deleteBrand($id);
+                header("Location: ".BASE_URL  . "admins");
+            }
+            else  $this->adminView->showMessage("No tienes permisos", 400);
+
         } catch (Exception $e) {
             $this->adminView->showMessage("La marca esta siendo utilizada por un producto",500);
         }
@@ -30,13 +35,17 @@ class BrandController
     public function new()
     {
         try{
-        
-            if (!isset($_POST['brand']) || !isset($_POST['id'])|| empty($_POST['brand'])) {
-                $this->adminView->showMessage("Bad request",400);              
-            } else {
-                $this->brandModel->add($_POST['brand']);
-                header("Location: ".BASE_URL  . "admins");
-            }           
+            session_start();     
+            if (isset($_SESSION['email']) && $_SESSION['role'] == UserRole::ADMIN)  {
+                if (!isset($_POST['brand']) || !isset($_POST['id'])|| empty($_POST['brand'])) {
+                    $this->adminView->showMessage("Bad request",400);              
+                } else {
+                    $this->brandModel->add($_POST['brand']);
+                    header("Location: ".BASE_URL  . "admins");
+                } 
+            }
+            else  $this->adminView->showMessage("No tienes permisos", 400);
+                      
         }catch(Exception $e){         
             
             $this->adminView->showMessage("La marca ingresada ya existe",500);
@@ -48,13 +57,18 @@ class BrandController
     public function edit()
     {  
         try{
-            if(isset($_POST['id']) && isset($_POST['brand'])&& !empty($_POST['brand'])) {       
-            $this->brandModel->edit($_POST['id'],$_POST['brand']);
-            header("Location: ".BASE_URL  . "admins");
+            session_start();     
+            if (isset($_SESSION['email']) && $_SESSION['role'] == UserRole::ADMIN)  {
+                if(isset($_POST['id']) && isset($_POST['brand'])&& !empty($_POST['brand'])) {       
+                    $this->brandModel->edit($_POST['id'],$_POST['brand']);
+                    header("Location: ".BASE_URL  . "admins");
+                }
+                else{
+                    $this->adminView->showMessage("Bad request",400);
+                }
             }
-            else{
-                $this->adminView->showMessage("Bad request",400);
-            }
+            else  $this->adminView->showMessage("No tienes permisos", 400);
+            
      
         }catch(Exception $e){
             $this->adminView->showMessage("La marca que se desea utilizar esta siendo utilizada por un producto",500);
